% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpm2_boot.R
\name{stmp2boot}
\alias{stmp2boot}
\title{Function to obtain CIF, difference in CIF, and Restricted Mean Time Lost (RMTL) from a stpm2 Stata model}
\usage{
stmp2boot(
  data,
  i,
  dat,
  stata_code,
  idvar,
  cif = F,
  diff = NULL,
  rmtl = NULL,
  group_names = c(),
  event = c("dementia"),
  ref = 1
)
}
\arguments{
\item{i}{Internal argument used by the boot function}

\item{dat}{The dataset to be analyzed}

\item{stata_code}{Stata code to run the stpm2 model and generate CIF predictions}

\item{idvar}{Character string specifying the ID variable in the dataset}

\item{cif}{Logical (TRUE or FALSE), indicating whether to return CIF estimates}

\item{diff}{Numeric vector of time points at which to compute differences in CIF (default is NULL; no differences are calculated)}

\item{rmtl}{Numeric vector of time horizons for calculating RMTL (default is NULL; RMTL is not calculated)}

\item{group_names}{Character vector specifying the names of the groups for CIF comparison}

\item{event}{The event of interest for CIF and RMTL calculations}

\item{ref}{Reference group for comparisons (default is 1)}
}
\description{
Function to obtain CIF, difference in CIF, and Restricted Mean Time Lost (RMTL) from a stpm2 Stata model
}
